requirements.txt
Flask==0.10.1
Jinja2==2.7.3
MarkupSafe==0.23
Werkzeug==0.9.6
argparse==1.2.1
gunicorn==19.1.1
itsdangerous==0.24
requests==2.5.3
wsgiref==0.1.2
billboard.py==4.2.0
